v1

+++ 01/megreSort.h
@@ -1,7 +1,7 @@
 #pragma once
 #include<iostream>
 
-void mergeSortedArr(int* arr, const size_t arrSize, int* arrLeft, const size_t arrLeftSize, int* arrRight, const size_t arrRightSize);
+void mergeSortedArr(int* arr, const int arrSize, int* arrLeft, const int arrLeftSize, int* arrRight, const int arrRightSize);
 
 
-void mergeSort(int* arr, const size_t sizeArr);
+void mergeSort(int* arr, const int sizeArr);
