+++ 02/quickSort.h
@@ -1,4 +1,4 @@
 #pragma once
 #include<iostream>
 
-void quickSort(int* arr, size_t arrLeft, size_t arrRight);
+void quickSort(int* arr, int arrLeft, int arrRight);
